requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
Pillow==9.1.0
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
